+++ digikambackend/src/main/java/org/braun/digikam/backend/model/Media.java
@@ -33,7 +33,7 @@ import jakarta.validation.constraints.*;
 public class Media   {
   public static final String JSON_PROPERTY_ID = "id";
   @JsonProperty(JSON_PROPERTY_ID)
-  private Integer id;
+  private Long id;
 
   public static final String JSON_PROPERTY_NAME = "name";
   @JsonProperty(JSON_PROPERTY_NAME)
@@ -47,7 +47,7 @@ public class Media   {
   @JsonProperty(JSON_PROPERTY_IMAGE)
   private Boolean image;
 
-  public Media id(Integer id) {
+  public Media id(Long id) {
     this.id = id;
     return this;
   }
@@ -59,11 +59,11 @@ public class Media   {
   @JsonProperty(value = "id")
   @ApiModelProperty(required = true, value = "")
   @NotNull 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
